Add a transaction to the budget from the context menu
Budgeting something that already exists meant reading the amount off a
transaction and typing it into the budget by hand.

Right-clicking a transaction, or a scheduled transaction, now offers Add
to Budget. The amount and category come from the source; the dialog asks
only for what the transaction cannot say by itself - which period and
segment it belongs to, how often it is expected, and whether to round it,
since a budget is a plan rather than a receipt.

A scheduled transaction suggests its own frequency from its repeat, and
can be linked rather than copied: choosing Automatic stores SCHED:<id> so
the budget keeps following the schedule instead of freezing a number that
will drift.

Budgeting the same category twice is a real possibility, so an existing
entry is never overwritten silently: the choice to replace, add to, or
leave it is explicit.

+++ src/model/PlanEngine.cpp
@@ -314,7 +314,7 @@ void PlanEngine::set_excluded_account_id_a(const std::vector<int64>& ids)
     for (const int64 id : ids) {
         if (!raw.IsEmpty())
             raw += ",";
-        raw += wxString::Format("%lld", static_cast<long long>(id.GetValue()));
+        raw += wxString::Format("%lld", id.GetValue());
     }
     InfoModel::instance().saveString(PLAN_EXCLUDED_ACCOUNTS, raw);
 }
